+++ activerecord/test/models/booking.rb
@@ -1,0 +1,4 @@
+class Booking < ActiveRecord::Base
+  has_many :booking_guests
+  belongs_to :booking_request
+end