+++ activerecord/test/models/booking_guest.rb
@@ -1,0 +1,4 @@
+class BookingGuest < ActiveRecord::Base
+  belongs_to :booking
+  belongs_to :bookable, polymorphic: true
+end