+++ activerecord/test/models/breakfast.rb
@@ -1,0 +1,3 @@
+class Breakfast < ActiveRecord::Base
+  has_many :booking_guests, as: :bookable
+end