+++ activerecord/test/models/guest.rb
@@ -1,0 +1,3 @@
+class Guest < ActiveRecord::Base
+  has_many :booking_guests, as: :bookable
+end